@@ -9,30 +9,33 @@
 		"build": "vite build",
 		"preview": "vite preview",
 		"check": "svelte-check",
-		"lint": "prettier --ignore-path .gitignore --check --plugin=prettier-plugin-svelte . && eslint --ignore-path .gitignore .",
+		"lint": "prettier --check --plugin=prettier-plugin-svelte . && eslint .",
 		"lint:css": "stylelint \"src/**/*.{css,svelte}\"",
 		"check:formatting": "prettier --check --plugin=prettier-plugin-svelte .",
-		"format": "prettier --ignore-path .gitignore --write --plugin=prettier-plugin-svelte ."
+		"format": "prettier --write --plugin=prettier-plugin-svelte ."
 	},
 	"devDependencies": {
-		"@commitlint/config-conventional": "^18.6.0",
-		"@fontsource/playfair-display": "^5.0.18",
-		"@rodneylab/sveltekit-components": "^3.0.0",
-		"@sveltejs/kit": "^2.5.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"commitlint": "^18.6.0",
-		"eslint": "^8.56.0",
+		"@commitlint/config-conventional": "^19.6.0",
+		"@eslint/js": "^9.17.0",
+		"@fontsource/playfair-display": "^5.1.1",
+		"@rodneylab/sveltekit-components": "^3.0.2",
+		"@sveltejs/kit": "^2.15.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"commitlint": "^19.6.1",
+		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"husky": "^9.0.10",
-		"postcss-html": "^1.6.0",
-		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.1.2",
-		"stylelint": "^16.2.1",
-		"stylelint-config-recommended": "^14.0.0",
-		"svelte": "^4.2.10",
-		"svelte-check": "^3.6.4",
-		"vite": "^5.1.0"
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.14.0",
+		"husky": "^9.1.7",
+		"postcss-html": "^1.7.0",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
+		"stylelint": "^16.12.0",
+		"stylelint-config-recommended": "^14.0.1",
+		"svelte": "^5.16.0",
+		"svelte-check": "^4.1.1",
+		"typescript-eslint": "^8.19.0",
+		"vite": "^6.0.6"
 	},
 	"type": "module",
 	"repository": {
